日記投稿機能の実装

+++ db/migrate/20201007082611_create_amounts.rb
@@ -1,7 +1,7 @@
 class CreateAmounts < ActiveRecord::Migration[5.2]
   def change
     create_table :amounts do |t|
-      t.float :amount, null: false
+      t.integer :amount, null: false
       t.integer :objective_id, null: false, foreign_key: true
       t.timestamps
     end
